requirements.txt
altair==4.1.0
aniso8601==9.0.1
APScheduler==3.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
backcall==0.2.0
backoff==1.11.1
backports.zoneinfo==0.2.1
bleach==4.1.0
bokeh==2.4.2
boltons==21.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-spinner==0.1.10
cmdstanpy==0.9.68
colorama==0.4.4
colorlog==6.6.0
colorlover==0.3.0
convertdate==2.3.2
cufflinks==0.17.3
cycler==0.11.0
Cython==0.29.26
dacite==1.6.0
datacommons==1.4.3
datacommons-pandas==0.0.3
datapane==0.13.0
DateTime==4.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dominate==2.6.0
entrypoints==0.3
ephem==4.1.3
face==20.1.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.28.5
furl==2.1.3
glom==20.11.0
gunicorn==20.1.0
hijri-converter==2.2.2
holidays==0.11.3.1
idna==3.3
importlib-resources==5.4.0
investpy==1.0.7
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
micawber==0.5.4
mistune==0.8.4
monotonic==1.6
munch==2.5.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.6
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
orderedmultidict==1.0.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
posthog==1.4.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pyarrow==5.0.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyMeeus==0.5.11
pyparsing==3.0.6
pyrsistent==0.18.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
setuptools-git==1.2
six==1.16.0
sklearn==0.0
stringcase==1.2.0
tabulate==0.8.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.0.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==4.4.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
Unidecode==1.3.2
urllib3==1.26.7
validators==0.18.2
vega-datasets==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wincertstore==0.2
wordcloud==1.8.1
zipp==3.6.0
zope.interface==5.4.0
